java 8 compatibility fixes

+++ sql/src/test/java/io/crate/integrationtests/InformationSchemaQueryTest.java
@@ -176,8 +176,8 @@ public class InformationSchemaQueryTest extends SQLTransportIntegrationTest {
     public void testNotEqualsNumber() throws Exception {
         exec("select table_name, number_of_shards from information_schema.tables where schema_name = 'doc' and number_of_shards != 7");
         assertEquals(2L, response.rowCount());
-        assertTrue(response.rows()[0][1] != 7);
-        assertTrue(response.rows()[1][1] != 7);
+        assertTrue((int)response.rows()[0][1] != 7);
+        assertTrue((int) response.rows()[1][1] != 7);
     }
 
     @Test
